moved calcualtor to servises
