+++ src/main/java/app/Main.java
@@ -8,9 +8,9 @@ import app.controllers.ItemController;
 import app.controllers.OrderController;
 import app.controllers.UserController;
 import app.controllers.PageController;
-import app.model.Calculator;
 import app.model.entities.Order;
 import app.persistence.ConnectionPool;
+import app.services.Calculator;
 import io.javalin.Javalin;
 import io.javalin.http.Context;
 import io.javalin.rendering.template.JavalinThymeleaf;
